# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Yadro API",
        "version": "1.0"
    },
    "paths": {
        "/account": {
            "get": {
                "tags": [
                    "account"
                ],
                "summary": "Возвращает информацию об аккаунте",
                "operationId": "7bc88bd17d6e87fa9a40c24918edfa63",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/account/statuses": {
            "get": {
                "tags": [
                    "account"
                ],
                "summary": "Возвращает информацию о статусах главной воронки",
                "operationId": "8e9d1f41691874b49926004aadef3c62",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/account/allStatuses": {
            "get": {
                "tags": [
                    "account"
                ],
                "summary": "Возвращает информацию о воронках и их статусах",
                "operationId": "51b5486a2cc571a2942730aa5183fa36",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/account/pipelines": {
            "get": {
                "tags": [
                    "account"
                ],
                "summary": "Возвращает информацию о воронках",
                "operationId": "088acd1eb0979e52d5d1ffb7c7a6f898",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/account/users": {
            "get": {
                "tags": [
                    "account"
                ],
                "summary": "Возвращает информацию о пользователях",
                "operationId": "3d340fa7df8434c1a2c417de133a4f48",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/account/fields": {
            "get": {
                "tags": [
                    "account"
                ],
                "summary": "Возвращает информацию о полях",
                "operationId": "0092baecda539f66c44cab0197cd870b",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/company/{id}": {
            "get": {
                "tags": [
                    "company"
                ],
                "summary": "Возвращает компанию по id",
                "operationId": "830ea207ae23d95b7e77de3d6cacd8b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id компании",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/company/search": {
            "get": {
                "tags": [
                    "company"
                ],
                "summary": "Поиск компаний",
                "operationId": "f2352c26adbed6f45104a9ec756bc7cc",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "строка поиска",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/company/list": {
            "get": {
                "tags": [
                    "company"
                ],
                "summary": "Возвращает список компаний",
                "operationId": "c6b4b9bedcf759825f765f18cda69cbd",
                "parameters": [
                    {
                        "name": "count",
                        "in": "query",
                        "description": "количество запрашиваемых элементов",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "смещение, относительно которого нужно вернуть элементы",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/company/create": {
            "post": {
                "tags": [
                    "company"
                ],
                "summary": "Создаёт компании",
                "operationId": "9135ad9a68c66ee292cb4e13fb9b2eba",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "example": {
                                        "name": "Company name"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/company/update": {
            "post": {
                "tags": [
                    "company"
                ],
                "summary": "Обновляет компании",
                "operationId": "7a55e72c1bd97c7cbcb6e4dbe22ed89a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "example": {
                                        "id": 12345678,
                                        "name": "New company name"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "account"
        },
        {
            "name": "bp"
        },
        {
            "name": "chat"
        },
        {
            "name": "company"
        },
        {
            "name": "contact"
        },
        {
            "name": "customer"
        },
        {
            "name": "docflow"
        },
        {
            "name": "field"
        },
        {
            "name": "lead"
        },
        {
            "name": "links"
        },
        {
            "name": "mail"
        },
        {
            "name": "note"
        },
        {
            "name": "pbx"
        },
        {
            "name": "sensei"
        },
        {
            "name": "sms"
        },
        {
            "name": "task"
        },
        {
            "name": "telegram"
        },
        {
            "name": "transaction"
        },
        {
            "name": "unsorted"
        },
        {
            "name": "yadro"
        }
    ],
    "security": [
        []
    ]
}